[data] Finalize two versions of PennTreebankDataset
The first version is an IterableDataset that stores the data as a stream of indices, and processes it to yield batches at run-time. The data pre-processing is limited to translating words to their indices in the vocabulary, and saving the stream of indices as a numpy array. Because the IterableDataset handles the batching, batch size must be set to None for the DataLoader.

The second version is a Dataset that stores a numpy array of pre-batched data. The preprocessing is more extensive here -- using the batch size and bptt length as parameters, the preprocessor pads the data and splits it into the correct batches for streaming, and then saves it as [input, target] pairs in an npy file. The data loader requests data one line at a time and handles the batching.