[analysis] Print x and alpha to log
This is a total hack -- we need x and alpha for heatmaps, so we print them to the log. To save them, pipe the output of a command to a file, and then press CTRL+C quickly so you don't create a huge final. Only one batch is needed for heat mapping. If anyone has a better way to do this, please help.